requirements.txt
# SHADOW Bug Bounty Automation Framework - Dependencies

# HTTP Client
aiohttp>=3.9.0
httpx>=0.25.0
requests>=2.31.0

# Async Support
asyncio-throttle>=1.0.0

# Configuration
pyyaml>=6.0

# Templating (Reports)
jinja2>=3.1.0

# HTML/XML Parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Browser Automation
playwright>=1.40.0

# DNS
dnspython>=2.4.0

# Cryptography & JWT
cryptography>=41.0.0
pyjwt>=2.8.0

# Date/Time
python-dateutil>=2.8.0

# CLI & Output
rich>=13.0.0
click>=8.1.0
tabulate>=0.9.0

# Data Validation
pydantic>=2.5.0

# URL Parsing
tldextract>=5.1.0
furl>=2.1.0

# Regex
regex>=2023.10.0

# Hashing
xxhash>=3.4.0

# Progress Bars
tqdm>=4.66.0

# Logging
loguru>=0.7.0

# File Operations
aiofiles>=23.2.0

# JSON Handling
orjson>=3.9.0

# Color Output
colorama>=0.4.6

# Testing (Development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
aioresponses>=0.7.0

# Type Checking (Development)
mypy>=1.7.0
types-requests>=2.31.0
types-pyyaml>=6.0.0

# Code Formatting (Development)
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
